requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
